% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionsMCMC.R
\name{proposeParamsMCMC}
\alias{proposeParamsMCMC}
\title{Sample from proposal distribution for MCMC}
\usage{
proposeParamsMCMC(oldParams, inferredParams, width)
}
\arguments{
\item{oldParams}{Previous parameter values}

\item{inferredParams}{Proposed parameter values}

\item{width}{Width of random walk proposal}
}
\value{
List with proposed parameters, indicator variable of which
parameters have changed, old and new proposal probabilities (if
}
\description{
Sample from proposal distribution for MCMC
}
