% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGP.R
\name{sigmoidVarKernCompute}
\alias{sigmoidVarKernCompute}
\title{Compute K(x, x2) for sigmoid kernel, used by gptk}
\usage{
sigmoidVarKernCompute(kern, x, x2 = NULL)
}
\arguments{
\item{kern}{kernel}

\item{x}{input 1}

\item{x2}{input 2}
}
\value{
K(x, x2)
}
\description{
Compute K(x, x2) for sigmoid kernel, used by gptk
}
