% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGP.R
\name{sigmoidVarKernDiagCompute}
\alias{sigmoidVarKernDiagCompute}
\title{Compute diagonal of sigmoid kernel (used by gptk).}
\usage{
sigmoidVarKernDiagCompute(kern, x)
}
\arguments{
\item{kern}{Kernel}

\item{x}{Input}
}
\value{
Diagonal of the kernel
}
\description{
Compute diagonal of sigmoid kernel (used by gptk).
}
