% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGP.R
\name{sigmoidVarKernExpandParam}
\alias{sigmoidVarKernExpandParam}
\title{Insert parameters into sigmoid kernel (used by gptk)}
\usage{
sigmoidVarKernExpandParam(kern, params)
}
\arguments{
\item{kern}{kernel}

\item{params}{parameters}
}
\value{
kernel
}
\description{
Insert parameters into sigmoid kernel (used by gptk)
}
