% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGP.R
\name{sigmoidVarKernExtractParam}
\alias{sigmoidVarKernExtractParam}
\title{Auxiliary function for sigmoid kernel (used by gptk)}
\usage{
sigmoidVarKernExtractParam(kern, only.values = TRUE,
  untransformed.values = TRUE)
}
\arguments{
\item{kern}{- kernel}

\item{only.values}{- return values only}

\item{untransformed.values}{- transform values}
}
\value{
hyperparameters of the GP kernel
}
\description{
Auxiliary function for sigmoid kernel (used by gptk)
}
