% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGP.R
\name{sigmoidVarKernGradient}
\alias{sigmoidVarKernGradient}
\title{Compute gradient of sigmoid kernel with respect to each parameter (used by gptk)}
\usage{
sigmoidVarKernGradient(kern, x, x2, covGrad)
}
\arguments{
\item{kern}{kernel}

\item{x}{input 1}

\item{x2}{input 2}

\item{covGrad}{gradient of covariance function}
}
\value{
d k(x, x2) / d theta
}
\description{
Compute gradient of sigmoid kernel with respect to each parameter (used by gptk)
}
