% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGP.R
\name{sigmoidVarKernParamInit}
\alias{sigmoidVarKernParamInit}
\title{Auxiliary function for sigmoid kernel (used by gptk)}
\usage{
sigmoidVarKernParamInit(kern)
}
\arguments{
\item{kern}{- GP kernel}
}
\value{
initialized kernel
}
\description{
Auxiliary function for sigmoid kernel (used by gptk)
}
