% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{solveODE}
\alias{solveODE}
\title{Solve ODE system explicitly.}
\usage{
solveODE(num.species, timePoints, ode.system, params)
}
\arguments{
\item{num.species}{Number of variables (species) in the system.}

\item{timePoints}{Time points at which to evaluate the ODE system.}

\item{ode.system}{Function for calculating the derivatives of the ODE system.}

\item{params}{Current values for the ODE parameter estimates.}
}
\value{
A list with two elements: \code{x} contains the results of integrating the ODE at the given time points, and \code{error} flags if there has been an error while invoking \code{deSolve}.
}
\description{
Solve ODE system explicitly.
}
