% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_mcmc.R
\name{doMCMC}
\alias{doMCMC}
\title{Main MCMC function
Runs the MCMC for the specified number of iterations and returns the 
sampled parameter values}
\usage{
doMCMC(timePoints, data, auxVars, options)
}
\arguments{
\item{timePoints}{Measured time points for the ODE system.}

\item{data}{Observed data.}

\item{auxVars}{Auxiliary variables.}

\item{options}{Options for MCMC run.}
}
\value{
List with samples parameters and information about the sampling run.
}
\description{
Main MCMC function
Runs the MCMC for the specified number of iterations and returns the 
sampled parameter values
}
